requirements.txt
image
pyzbar~=0.1.8
Pillow~=8.1.0